requirements.txt
fastapi
uvicorn
pillow
telethon
python-dotenv
